fix: refactor clear icon to button element to make it accessible
