+++ assets/index.less
@@ -411,7 +411,7 @@
   &-clear {
     position: absolute;
     top: 0;
-    inset-inline-end: 4px;
+    inset-inline-end: 0;
     cursor: pointer;
 
     &-btn::after {
