+++ src/PickerInput/Selector/Icon.tsx
@@ -4,41 +4,18 @@ import { clsx } from 'clsx';
 
 export interface IconProps extends React.HtmlHTMLAttributes<HTMLElement> {
   icon?: React.ReactNode;
-  type: 'suffix' | 'clear';
 }
 
-export default function Icon(props: IconProps) {
-  const { icon, type, ...restProps } = props;
+export default function Icon({ icon, ...restProps }: IconProps) {
   const { prefixCls, classNames, styles } = React.useContext(PickerContext);
 
   return icon ? (
     <span
-      className={clsx(`${prefixCls}-${type}`, classNames.suffix)}
+      className={clsx(`${prefixCls}-suffix`, classNames.suffix)}
       style={styles.suffix}
       {...restProps}
     >
       {icon}
     </span>
   ) : null;
-}
-
-export interface ClearIconProps extends Omit<IconProps, 'type'> {
-  onClear: VoidFunction;
-}
-
-export function ClearIcon({ onClear, ...restProps }: ClearIconProps) {
-  return (
-    <Icon
-      {...restProps}
-      type="clear"
-      role="button"
-      onMouseDown={(e) => {
-        e.preventDefault();
-      }}
-      onClick={(e) => {
-        e.stopPropagation();
-        onClear();
-      }}
-    />
-  );
 }
